requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-debug-toolbar==4.4.6
google-auth==2.37.0
python-dotenv~=1.0.1